Manager can now register employees
