+++ Backend/src/main/java/com/happytravel/happytravel/api/controller/UserController.java
@@ -21,11 +21,8 @@ public class UserController {
 
     private final UserService userService;
     private final PersonService personService;
-    private final EmployeeService employeeService;
     private final ClientService clientService;
-    private final ManagerService managerService;
-    private final GuideService guideService;
-    private final SellerService sellerService;
+    private final EmployeeService employeeService;
 
     @GetMapping("/users")
     @ResponseStatus(HttpStatus.OK)
@@ -41,8 +38,20 @@ public class UserController {
         if (user == null) {
             result = -1;
             return "{\"result\":" + result + ", \"user\": null}";
-        } else result = 0;
+        }
+        if(!userService.getUserType(login).equals("Klient")){
+            //return "Sprawdzam date";
+            if(employeeService.getFiringDateByUserID(userService.getUserIdByLogin(login))!=null){
+                result = -1;
+                return "Zwolniony";
+                //return "{\"result\":" + result + ", \"user\": null}";
+            }
+        }
+        result = 0;
         return "{\"result\":" + result + ", \"user\":" + user.toString() + "}";
+        //result = -1;
+        //return "aaa";
+        //return "{\"result\":" + result + ", \"user\": null}";
     }
 
     @GetMapping("/signup")
@@ -63,51 +72,5 @@ public class UserController {
             return 0;
         }
     }
-    @GetMapping("/signEmployee")
-    public String signNewEmployee(@RequestParam Long loggedUser, @RequestParam String name, @RequestParam String lastName, @RequestParam String email, @RequestParam Long pesel, @RequestParam String phoneNumber, @RequestParam String type, @RequestParam Long salary, @RequestParam Long officeID){
-        if(!isManager(loggedUser)) return "-1";
-        else if(personService.getPersonByEmail(email) != null) return "-2";
-        else {
-            Long userid = userService.getMaxId() + 1;
-            int result = 0;
-            String login = type + userid;
-            String alphabet = "qwertyuiopasdfghjklzxcvbnm";
-            String pass = "";
-            Random rand = new Random();
-            for(int i=0;i<8;i++){
-                pass+=alphabet.charAt(rand.nextInt(alphabet.length()));
-            }
-            result = userService.signUp(userid, login, pass, type);
-            if (result == 0) return "-3";
-            Long personId = personService.getMaxId() + 1;
-            result = personService.insertPerson(personId, email, name, pesel, phoneNumber, lastName, userid);
-            if (result == 0) return "-4";
-            Long employeeId = employeeService.getMaxId()+1;
-            result = employeeService.insertEmployee(employeeId, null, new Date(), type, salary, officeID, personId);
-            if (result == 0) return "-5";
-            if (type.equals("Kierownik")){
-                Long managerId = managerService.getMaxId()+1;
-                result = managerService.insertManager(managerId, employeeId);
-                if(result == 0) return "-6";
-            }
-            else if (type.equals("Przewodnik")){
-                Long guideId = guideService.getMaxId()+1;
-                result = guideService.insertGuide(guideId, employeeId);
-                if(result == 0) return "-7";
-            }
-            else if (type.equals("Sprzedawca")){
-                Long sellerId = sellerService.getMaxId()+1;
-                result = sellerService.insertSeller(sellerId, employeeId);
-                if(result == 0) return "-8";
-            }
-            User user = userService.getUserByID(userid);
-            if(user == null) return "-9";
-            return "{\"user\":" + user.toString() + "}";
-        }
-        }
-    @GetMapping("/isManager")
-    public boolean isManager(@RequestParam Long user_ID){
-        if(userService.getManagerID(user_ID)==null) return false;
-        else return true;
-    }
+
 }
