+++ Backend/src/main/java/com/happytravel/happytravel/api/dto/OfficeDto.java
@@ -8,5 +8,5 @@ import lombok.Data;
 @Data
 public class OfficeDto {
 	private Long id;
-	private Adres addressID;
+	private Adres adressID;
 }
