+++ Backend/src/main/java/com/happytravel/happytravel/api/repository/EmployeeRepository.java
@@ -22,4 +22,6 @@ public interface EmployeeRepository extends JpaRepository<Employee, Long> {
     int insertEmployee(@Param("id") Long id, @Param("firingdate") Date firingdate, @Param("hiringdate") Date hiringdate, @Param("position") String position, @Param("salary") Long salary, @Param("officeid") Long officeid, @Param("personid") Long personid);
     @Query(value = "select max(id) from Employee", nativeQuery = true)
     Long getMaxId();
+    @Query(value = "select distinct Employee.firingDate from Employee, User, Person where Employee.personID = Person.id and :userID = Person.userID", nativeQuery=true)
+    Date getFiringDateByUserID(@Param("userID") Long userID);
 }
