+++ Backend/src/main/java/com/happytravel/happytravel/api/repository/ManagerRepository.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Date;
 
 @Repository
 public interface ManagerRepository extends JpaRepository<Manager, Long> {
@@ -23,5 +24,11 @@ public interface ManagerRepository extends JpaRepository<Manager, Long> {
     int insertManager(@Param("id") Long id, @Param("employeeId") Long employeeId);
     @Query(value = "Select * from Manager where id = :mID", nativeQuery = true)
     Manager getManagerByID(@Param("mID") Long mID);
+    @Modifying
+    @Transactional
+    @Query(value = "update Employee set firingdate = :date where id = :eID", nativeQuery = true)
+    int fireEmployee(@Param("eID") Long eID, @Param("date") Date date);
+    @Query(value = "SELECT Manager.id from Manager, Employee , Person  WHERE Employee.id = Manager.EmployeeID AND Person.id = Employee.personID AND Person.userID = :user_ID", nativeQuery = true)
+    Long getManagerID(@Param("user_ID") Long user_ID);
 
 }
