+++ Backend/src/main/java/com/happytravel/happytravel/api/service/EmployeeService.java
@@ -10,5 +10,6 @@ public interface EmployeeService {
     List<Employee> getEmployees();
     int insertEmployee(Long id, Date firingdate, Date hiringdate, String position, Long salary, Long officeid, Long personid);
     Long getMaxId();
+    Date getFiringDateByUserID(Long userID);
 
 }
