+++ Backend/src/main/java/com/happytravel/happytravel/api/service/ManagerService.java
@@ -3,6 +3,7 @@ package com.happytravel.happytravel.api.service;
 import com.happytravel.happytravel.api.model.Manager;
 
 import java.util.List;
+import java.util.Date;
 
 public interface ManagerService {
 
@@ -10,4 +11,6 @@ public interface ManagerService {
     Long getMaxId();
     int insertManager(Long id, Long employeeId);
     Manager getManagerByID(Long mID);
+    int fireEmployee(Long eID, Date date);
+    Long getManagerID(Long user_ID);
 }
